删除管理员时进行判断，是否能被删除

+++ src/main/java/com/nf/service/impl/AdministratorsServiceImpl.java
@@ -52,6 +52,10 @@ public class AdministratorsServiceImpl implements AdministratorsService {
 
     @Override
     public boolean delete(int id) {
+        AdministratorsEntity admin = dao.getAdmin(id);
+        if (admin.getAuthority()){
+            return false;
+        }
         return 1 == dao.delete(id);
     }
 }
